# docs/swagger.yaml
definitions:
  advise.Notification:
    properties:
      content:
        description: 通知内容，富文本
        minLength: 2
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      expired:
        description: 过期时间 2023-05-24 17:01:11
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  bbs.BBS:
    properties:
      bbsContent:
        allOf:
        - $ref: '#/definitions/creeper.Content'
        description: 论坛信息
      userInfo:
        allOf:
        - $ref: '#/definitions/user.User'
        description: 用户信息
    type: object
  common.BaseResponse:
    properties:
      code:
        description: 1 成功 0 失败
        type: integer
      content:
        description: 'in: body'
      message:
        description: 错误信息
        type: string
    type: object
  creeper.Content:
    properties:
      conditions:
        description: 条件 让球 1.25
        items:
          type: string
        type: array
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      extra:
        description: 额外的一些信息
        type: string
      id:
        type: integer
      imageUrl:
        items:
          type: string
        type: array
      match:
        description: 比赛
        type: string
      predict:
        description: 预测谁赢
        type: string
      summery:
        type: string
      title:
        type: string
      type:
        description: 网站名称
        example: 雷速
        type: string
      updatedAt:
        type: string
      url:
        type: string
      userID:
        type: integer
    type: object
  order.LotteryDetail:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      odds:
        description: 赔率
        type: number
      parentId:
        type: integer
      scoreVsScore:
        description: |-
          比分， 类型BF才有 s00s00 s05s02
          半全场胜平负， 类型BQSFP  aa hh
          总进球数， 类型ZJQ s0 - s7
          胜负平， 类型SFP hada主负 hadd主平 hadh 主胜  hhada客负 hhadd客平 hhadh 客胜
        type: string
      type:
        description: 类型 枚举：SFP（胜负平）、BF（比分）、ZJQ(总进球)、BQSFP（半全场胜负平）
        type: string
      updatedAt:
        type: string
    type: object
  order.Match:
    properties:
      awayTeamAllName:
        description: 客队全名
        type: string
      awayTeamCode:
        description: 客队缩写
        type: string
      awayTeamId:
        description: 客队id
        type: string
      awayTeamName:
        description: 客队名称
        type: string
      combines:
        description: 彩票组合
        items:
          $ref: '#/definitions/order.LotteryDetail'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      homeTeamAllName:
        description: 主队全名
        type: string
      homeTeamCode:
        description: 主队缩写
        type: string
      homeTeamId:
        description: 主队id
        type: string
      homeTeamName:
        description: 主队名称
        type: string
      id:
        type: integer
      leagueAllName:
        description: 联赛全名
        type: string
      leagueCode:
        description: 联赛编号
        type: string
      leagueId:
        description: 联赛id
        type: string
      leagueName:
        description: 联赛名称
        type: string
      matchDate:
        description: 比赛时间 2023-05-23 01:10:00
        type: string
      matchId:
        description: 比赛id
        type: string
      matchNum:
        description: 比赛编号
        type: string
      orderId:
        type: string
      updatedAt:
        type: string
    required:
    - awayTeamAllName
    - awayTeamCode
    - awayTeamId
    - awayTeamName
    - combines
    - homeTeamAllName
    - homeTeamCode
    - homeTeamId
    - homeTeamName
    - leagueAllName
    - leagueCode
    - leagueId
    - leagueName
    - matchDate
    - matchId
    - matchNum
    - orderId
    type: object
  order.Order:
    properties:
      allWinId:
        description: 合买id
        type: integer
      content:
        description: 数字内容
        type: string
      logicWinMaX:
        description: 逻辑最大中奖
        type: number
      logicWinMin:
        description: 逻辑最小中奖
        type: number
      lotteryType:
        description: 足彩（FOOTBALL） 大乐透（SUPER_LOTTO）  排列三（P3） 篮球(BASKETBALL) 七星彩（SEVEN_STAR）
          排列五（P5）
        type: string
      lotteryUuid:
        description: 查询id
        type: string
      matches:
        description: 比赛
        items:
          $ref: '#/definitions/order.Match'
        type: array
      saveType:
        description: 保存类型 TEMP（临时保存） TOMASTER（提交到店）  合买(ALLWIN)
        type: string
      share:
        description: 是否让人跟单
        type: boolean
      times:
        description: 倍数
        type: integer
      userID:
        description: 用户编号
        type: integer
      uuid:
        type: string
      way:
        description: 过关
        type: string
    required:
    - times
    type: object
  user.Complain:
    properties:
      comment:
        description: 修改备注
        type: string
      content:
        description: 投诉详情
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      image:
        description: 图片
        type: string
      phone:
        description: 联系电话
        type: string
      proposal:
        description: 建议
        type: string
      type:
        description: 投诉类型
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
      userName:
        type: string
    type: object
  user.TokenVO:
    properties:
      token:
        type: string
    type: object
  user.User:
    properties:
      ali:
        description: 支付宝号
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        description: 昵称
        type: string
      phone:
        type: string
      role:
        description: '"enum: Admin,User"'
        type: string
      salt:
        description: 盐
        type: string
      score:
        description: 余额
        type: number
      secret:
        description: 密码
        type: string
      updatedAt:
        type: string
      wechat:
        description: 微信号
        type: string
    type: object
  user.UserDTO:
    properties:
      ali:
        description: 支付宝号
        type: string
      name:
        description: 昵称
        type: string
      phone:
        type: string
      role:
        description: '"enum: Admin,User"'
        type: string
      score:
        description: 余额
        type: number
      wechat:
        description: 微信号
        type: string
    type: object
  user.UserVO:
    properties:
      name:
        description: 昵称
        maxLength: 16
        minLength: 4
        type: string
      phone:
        maxLength: 11
        minLength: 11
        type: string
      secret:
        description: 密码
        maxLength: 16
        minLength: 6
        type: string
    type: object
info:
  contact: {}
paths:
  /bbs/commit:
    post:
      consumes:
      - application/json
      description: 提交论坛的帖子
      parameters:
      - description: 提交对象
        in: body
        name: param
        schema:
          $ref: '#/definitions/bbs.BBS'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 提交帖子
  /bbs/list:
    get:
      consumes:
      - application/json
      description: 提交论坛的帖子
      parameters:
      - description: 日期 unix time
        in: query
        name: date
        type: integer
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 查询全部贴子
  /cache:
    post:
      consumes:
      - application/json
      description: 内部缓存使用接口
      parameters:
      - description: 提交对象
        in: body
        name: param
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 内部缓存使用接口
  /download:
    get:
      consumes:
      - application/json
      description: 下载文件/图片
      parameters:
      - description: 文件名称
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 下载文件/图片
  /notify:
    get:
      consumes:
      - application/json
      description: 查询通告
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 查询通告
  /order:
    post:
      consumes:
      - application/json
      description: 订单创建接口
      parameters:
      - description: 订单对象
        in: body
        name: param
        schema:
          $ref: '#/definitions/order.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 订单创建接口
  /ping:
    get:
      consumes:
      - application/json
      description: 状态检测
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
  /salt:
    get:
      consumes:
      - application/json
      description: 公钥 默认10分钟过期
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 获取加密的公钥
  /super/complains:
    get:
      consumes:
      - application/json
      description: 查看投诉
      parameters:
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 查看投诉
  /super/creep:
    get:
      consumes:
      - application/json
      description: 爬虫接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 爬虫接口
  /super/notify:
    post:
      consumes:
      - application/json
      description: 创建通告
      parameters:
      - description: 通告对象
        in: body
        name: param
        schema:
          $ref: '#/definitions/advise.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 创建通告
  /upload:
    post:
      consumes:
      - application/json
      description: 上传文件或者图片
      parameters:
      - description: BBS,USER,HEADER 或则null
        in: query
        name: type
        type: string
      - description: 文件
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 上传文件或者图片
  /user:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 用户对象
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/user.UserVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 创建用户
  /user/complain:
    post:
      consumes:
      - application/json
      description: 投诉
      parameters:
      - description: 投诉对象
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/user.Complain'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 投诉
  /user/info:
    get:
      consumes:
      - application/json
      description: 查询用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 查询用户信息
    post:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户对象, socre 可以不传
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/user.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 更新用户信息
  /user/login:
    post:
      consumes:
      - application/json
      description: 公钥放在头里 salt， 密码：需要和公钥rsa 加密 账号为手机号
      parameters:
      - description: 用户对象
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/user.UserVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 登录接口
  /user/logout:
    post:
      consumes:
      - application/json
      description: 注销当前用户
      parameters:
      - description: token对象
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/user.TokenVO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      summary: 注销登录
swagger: "2.0"